improved "hil"
It's completely reworked.  The previous version had a fundamental flaw, in that it allowed the
second (and third and fourth ...) match iterations to match against the ANSI escape code that was
generated in earlier iterations.  This sometimes meant that ANSI codes got mangled + nested, in a
way that was very non-obvious to the person using hil.

This version solves that problem by marching monotonically through the string, rather than
restarting the match at the beginning of the string every single time.