[HTML template] make <td> in header rows invisible
Sometimes I'll intermix <td> and <th> in header rows.  Usually, I have a faint border around every
table cell, but for <td>s in the header rows, it looks out of place.

To trigger this border-hiding behavior, use class="header" on <tr>s that are header rows, or wrap
the rows in <thead>.