[rmrf] bail immediately if no files are found
This is often because the user misspelled a filename.  Before this change, they had to go through
the rigamarole of answering "yes" or "no", maybe misspelling THAT, and getting stuck for a second,
when they already know that they had a typo, and just want to get back to the cmdline and fix it.